+++ supvisors/tests/test_mapper.py
@@ -17,6 +17,8 @@
 # limitations under the License.
 # ======================================================================
 
+from unittest.mock import call
+
 import pytest
 
 from supvisors.internal_com.mapper import *
@@ -49,6 +51,7 @@ def test_sup_id_create_no_match(supvisors):
         assert sup_id.aliases is None
         assert sup_id.ip_addresses is None
         assert sup_id.ip_address is None
+        assert sup_id.stereotypes == []
         with pytest.raises(TypeError):
             str(sup_id)
 
@@ -68,6 +71,7 @@ def test_sup_id_create_simple_no_default(supvisors):
     assert sup_id.aliases is None
     assert sup_id.ip_addresses is None
     assert sup_id.ip_address is None
+    assert sup_id.stereotypes == []
     with pytest.raises(TypeError):
         str(sup_id)
 
@@ -85,6 +89,7 @@ def test_sup_id_create_host(supvisors):
     assert sup_id.aliases == ['10.0.0.1']
     assert sup_id.ip_addresses == ['10.0.0.1']
     assert sup_id.ip_address == '10.0.0.1'
+    assert sup_id.stereotypes == []
     assert str(sup_id) == '10.0.0.1'
 
 
@@ -101,6 +106,7 @@ def test_sup_id_create_host_port(supvisors):
     assert sup_id.aliases == ['10.0.0.1']
     assert sup_id.ip_addresses == ['10.0.0.1']
     assert sup_id.ip_address == '10.0.0.1'
+    assert sup_id.stereotypes == []
     assert str(sup_id) == '10.0.0.1:7777'
     # test with host+ports match (internal port defined)
     supvisors.options.event_port = 0
@@ -114,6 +120,7 @@ def test_sup_id_create_host_port(supvisors):
     assert sup_id.ip_addresses == ['10.0.0.1']
     assert sup_id.ip_address == '10.0.0.1'
     assert sup_id.event_port == 8001  # defaulted to internal_port + 1
+    assert sup_id.stereotypes == []
     assert str(sup_id) == '10.0.0.1:7777'
     # test with host+ports match (internal port defined before HTTP port)
     supvisors.options.event_port = 0
@@ -127,6 +134,7 @@ def test_sup_id_create_host_port(supvisors):
     assert sup_id.aliases == ['10.0.0.1']
     assert sup_id.ip_addresses == ['10.0.0.1']
     assert sup_id.ip_address == '10.0.0.1'
+    assert sup_id.stereotypes == []
     assert str(sup_id) == '10.0.0.1:7777'
 
 
@@ -143,6 +151,7 @@ def test_sup_id_create_identifier(supvisors):
     assert sup_id.aliases == ['cliche81']
     assert sup_id.ip_addresses == ['cliche81']
     assert sup_id.ip_address == 'cliche81'
+    assert sup_id.stereotypes == []
     # test with identifier set
     sup_id = SupvisorsInstanceId('<supvisors>cliche81:8888:5555', supvisors)
     assert sup_id.identifier == 'supvisors'
@@ -154,6 +163,7 @@ def test_sup_id_create_identifier(supvisors):
     assert sup_id.aliases == ['cliche81']
     assert sup_id.ip_addresses == ['cliche81']
     assert sup_id.ip_address == 'cliche81'
+    assert sup_id.stereotypes == []
 
 
 def test_sup_id_host_matches(supvisors):
@@ -215,18 +225,18 @@ def test_mapper_configure(mocker, mapper):
     # configure mapper with elements
     items = ['127.0.0.1', '<supervisor_05>10.0.0.5:7777:', '10.0.0.4:15000:8888', '10.0.0.5:9999:']
     core_items = ['127.0.0.1', 'supervisor_05', '10.0.0.4:15000', 'supervisor_06', '10.0.0.4']
-    mapper.configure(items, core_items)
+    mapper.configure(items, 'test', core_items)
     assert list(mapper.instances.keys()) == ['127.0.0.1', 'supervisor_05', '10.0.0.4:15000', '10.0.0.5:9999']
     assert mapper.core_identifiers == ['127.0.0.1', 'supervisor_05', '10.0.0.4:15000']
     assert mapper.nodes == {'127.0.0.1': ['127.0.0.1'], '10.0.0.5': ['supervisor_05', '10.0.0.5:9999'],
                             '10.0.0.4': ['10.0.0.4:15000']}
     assert mapper.initial_identifiers == ['127.0.0.1', 'supervisor_05', '10.0.0.4:15000', '10.0.0.5:9999']
-    assert mocked_find.called
+    assert mocked_find.call_args_list == [call('test')]
     mocked_find.reset_mock()
     # configure mapper with one invalid element
     items = ['127.0.0.1', '<dummy_id>', '<supervisor_05>10.0.0.5', '10.0.0.4:15000']
     with pytest.raises(ValueError):
-        mapper.configure(items, core_items)
+        mapper.configure(items, '', core_items)
     assert not mocked_find.called
 
 
@@ -237,19 +247,24 @@ def test_find_local_identifier_fqdn(mapper):
     for item in items:
         supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
         mapper._instances[supvisors_id.identifier] = supvisors_id
+        assert supvisors_id.stereotypes == []
     # 1. force host name and aliases
     sup_id = mapper._instances[fqdn]
     sup_id.host_name = fqdn
     sup_id.aliases = ['dummy']
     # find self
-    mapper.find_local_identifier()
+    mapper.find_local_identifier({'test'})
     assert mapper.local_identifier == fqdn
+    for identifier, supvisors_id in mapper.instances.items():
+        assert identifier != fqdn or supvisors_id.stereotypes == ['test']
     # 2. move fqdn
     sup_id.host_name = 'dummy'
     sup_id.aliases = [fqdn]
     # find self
-    mapper.find_local_identifier()
+    mapper.find_local_identifier({'test_2'})
     assert mapper.local_identifier == fqdn
+    for identifier, supvisors_id in mapper.instances.items():
+        assert identifier != fqdn or supvisors_id.stereotypes == ['test_2']
 
 
 def test_find_local_identifier_host_name(mapper):
@@ -260,19 +275,24 @@ def test_find_local_identifier_host_name(mapper):
     for item in items:
         supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
         mapper._instances[supvisors_id.identifier] = supvisors_id
+        assert supvisors_id.stereotypes == []
     # 1. force host name and aliases
     sup_id = mapper._instances[f'{host_name}:65000']
     sup_id.host_name = fqdn
     sup_id.aliases = ['dummy']
     # find self
-    mapper.find_local_identifier()
+    mapper.find_local_identifier({'test'})
     assert mapper.local_identifier == f'{host_name}:65000'
+    for identifier, supvisors_id in mapper.instances.items():
+        assert identifier != fqdn or supvisors_id.stereotype == ['test']
     # 2. move fqdn
     sup_id.host_name = 'dummy'
     sup_id.aliases = [fqdn]
     # find self
-    mapper.find_local_identifier()
+    mapper.find_local_identifier({'test_2'})
     assert mapper.local_identifier == f'{host_name}:65000'
+    for identifier, supvisors_id in mapper.instances.items():
+        assert identifier != fqdn or supvisors_id.stereotype == ['test_2']
 
 
 def test_find_local_identifier_ip_address(mapper):
@@ -284,19 +304,24 @@ def test_find_local_identifier_ip_address(mapper):
     for item in items:
         supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
         mapper._instances[supvisors_id.identifier] = supvisors_id
+        assert supvisors_id.stereotypes == []
     # 1. force host name and aliases
     sup_id = mapper._instances[f'host']
     sup_id.host_name = fqdn
     sup_id.aliases = ['dummy']
     # find self
-    mapper.find_local_identifier()
+    mapper.find_local_identifier({'test'})
     assert mapper.local_identifier == 'host'
+    for identifier, supvisors_id in mapper.instances.items():
+        assert identifier != fqdn or supvisors_id.stereotypes == ['test']
     # 2. move fqdn
     sup_id.host_name = 'dummy'
     sup_id.aliases = [fqdn]
     # find self
-    mapper.find_local_identifier()
+    mapper.find_local_identifier({'test_2'})
     assert mapper.local_identifier == 'host'
+    for identifier, supvisors_id in mapper.instances.items():
+        assert identifier != fqdn or supvisors_id.stereotypes == ['test_2']
 
 
 def test_find_local_identifier_inconsistent(mapper):
@@ -305,10 +330,12 @@ def test_find_local_identifier_inconsistent(mapper):
     host_name = gethostname()
     supvisors_id = SupvisorsInstanceId(f'<{host_name}>10.0.0.5:7777:', mapper.supvisors)
     mapper._instances[supvisors_id.identifier] = supvisors_id
+    assert supvisors_id.stereotypes == []
     # find self
     print(mapper.supvisors.supervisor_data.identifier)
     with pytest.raises(ValueError):
-        mapper.find_local_identifier()
+        mapper.find_local_identifier({'test'})
+    assert supvisors_id.stereotypes == []
 
 
 def test_find_local_identifier_multiple(mapper):
@@ -320,11 +347,14 @@ def test_find_local_identifier_multiple(mapper):
     for item in items:
         supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
         mapper._instances[supvisors_id.identifier] = supvisors_id
+        assert supvisors_id.stereotypes == []
         # update default processing of host
         supvisors_id.host_name = fqdn
     # find self
     with pytest.raises(ValueError):
-        mapper.find_local_identifier()
+        mapper.find_local_identifier({'test'})
+    for supvisors_id in mapper._instances.values():
+        assert supvisors_id.stereotypes == []
 
 
 def test_find_local_identifier_none(mapper):
@@ -336,19 +366,67 @@ def test_find_local_identifier_none(mapper):
         supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
         mapper._instances[supvisors_id.identifier] = supvisors_id
     with pytest.raises(ValueError):
-        mapper.find_local_identifier()
+        mapper.find_local_identifier({'test'})
+    for supvisors_id in mapper._instances.values():
+        assert supvisors_id.stereotypes == []
+
+
+def test_assign_stereotypes(mapper):
+    """ Test the SupvisorsMapper.assign_stereotypes method. """
+    # add context
+    items = ['10.0.0.1', '10.0.0.3', '10.0.0.4', '10.0.0.2']
+    for item in items:
+        supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
+        mapper._instances[supvisors_id.identifier] = supvisors_id
+        assert supvisors_id.stereotypes == []
+    assert mapper.stereotypes == {}
+    # test with empty stereotype
+    mapper.assign_stereotypes('10.0.0.1', set())
+    assert mapper.stereotypes == {}
+    assert all(x.stereotypes == [] for x in mapper.instances.values())
+    # assign stereotype to a Supvisors instance
+    mapper.assign_stereotypes('10.0.0.3', {'test'})
+    assert mapper.stereotypes == {'test': ['10.0.0.3']}
+    assert mapper.instances['10.0.0.1'].stereotypes == []
+    assert mapper.instances['10.0.0.2'].stereotypes == []
+    assert mapper.instances['10.0.0.3'].stereotypes == ['test']
+    assert mapper.instances['10.0.0.4'].stereotypes == []
+    # assign same stereotype to another Supvisors instance that is before in the instances order
+    mapper.assign_stereotypes('10.0.0.1', {'test'})
+    assert mapper.stereotypes == {'test': ['10.0.0.1', '10.0.0.3']}
+    assert mapper.instances['10.0.0.1'].stereotypes == ['test']
+    assert mapper.instances['10.0.0.2'].stereotypes == []
+    assert mapper.instances['10.0.0.3'].stereotypes == ['test']
+    assert mapper.instances['10.0.0.4'].stereotypes == []
+    # assign same stereotype to another Supvisors instance that is after in the instances order
+    mapper.assign_stereotypes('10.0.0.2', {'test', 'other test'})
+    assert mapper.stereotypes == {'test': ['10.0.0.1', '10.0.0.3', '10.0.0.2'], 'other test': ['10.0.0.2']}
+    assert mapper.instances['10.0.0.1'].stereotypes == ['test']
+    assert sorted(mapper.instances['10.0.0.2'].stereotypes) == sorted(['test', 'other test'])
+    assert mapper.instances['10.0.0.3'].stereotypes == ['test']
+    assert mapper.instances['10.0.0.4'].stereotypes == []
+    # assign another stereotype to another Supvisors instance
+    mapper.assign_stereotypes('10.0.0.4', {'other test'})
+    assert mapper.stereotypes == {'test': ['10.0.0.1', '10.0.0.3', '10.0.0.2'], 'other test': ['10.0.0.4', '10.0.0.2']}
+    assert mapper.instances['10.0.0.1'].stereotypes == ['test']
+    assert sorted(mapper.instances['10.0.0.2'].stereotypes) == sorted(['test', 'other test'])
+    assert mapper.instances['10.0.0.3'].stereotypes == ['test']
+    assert mapper.instances['10.0.0.4'].stereotypes == ['other test']
 
 
 def test_filter(mapper):
     """ Test the SupvisorsMapper.filter method. """
     # add context
     hostname = gethostname()
-    items = ['127.0.0.1', '<host>10.0.0.1:2222:', f'{hostname}:60000:7777']
+    items = ['127.0.0.1', '<host>10.0.0.1:2222:', f'{hostname}:60000:7777', '10.0.0.2']
     for item in items:
         supvisors_id = SupvisorsInstanceId(item, mapper.supvisors)
         mapper._instances[supvisors_id.identifier] = supvisors_id
     # force host name to FQDN
     mapper._instances[f'{hostname}:60000'].host_name = getfqdn()
     # test with a bunch of identifiers
-    identifier_list = ['127.0.0.1', 'host', f'{hostname}:60000', hostname, 'host', 'supervisor', '10.0.0.1']
-    assert mapper.filter(identifier_list) == ['127.0.0.1', 'host', f'{hostname}:60000']
+    identifier_list = ['host', f'{hostname}:60000', hostname, 'host', 'supervisor', '10.0.0.1']
+    assert mapper.filter(identifier_list) == ['host', f'{hostname}:60000']
+    # set stereotype to one of them
+    mapper.assign_stereotypes('10.0.0.2', {'supervisor'})
+    assert mapper.filter(identifier_list) == ['host', f'{hostname}:60000', '10.0.0.2']